remove timestamp from dne, add option to include it
Recording the timestamp in generated files creates unnecessary noise as
files that haven't really changed at all will show up as different when
comparing builds, which is something downstream distributions would like
to do.

default to not adding date and version information (old behaviour when
'-d' was passed) and add a new option '-D' that restores the timestamp,
in case anyone desperately needs it.