dne - Do Not Edit and AUTOGEN_DNE_DATE
This Scheme function has "traditionally" included a date and time stamp
in the resulting output and that could be suppressed by providing it a
"-d" option.  Due to very reasonable complaints, the new default is to
<strong>not</strong> include the date and require a "-D" option to
<i>en</i>able dates.  However, even that is not completely satisfactory.
The <strong>user</strong> of the template needs a way to select
behavior.  Enter the newly recognized environment variable:
<tt>AUTOGEN_DNE_DATE</tt> If set, it will override any "-D" option in
the template.  If set to the empty string or a string with the first
character being one of <tt>0nNfF</tt> (zero or the first letter of
"false" or "no"), then the date will not be emitted.  Otherwise, the
presence of this environment variable will cause the date to be emitted.